ADDED COMMAND SYSTEM THAT ACTUALLY WORKS?????? ez fix snelle zelfbouw

+++ 3rdParty/SDL2/include/SDL_mouse.h
@@ -443,11 +443,11 @@ extern DECLSPEC int SDLCALL SDL_ShowCursor(int toggle);
  * - Button 3:  Right mouse button
  */
 #define SDL_BUTTON(X)       (1 << ((X)-1))
-#define SDL_BUTTON_LEFT     1
-#define SDL_BUTTON_MIDDLE   2
-#define SDL_BUTTON_RIGHT    3
-#define SDL_BUTTON_X1       4
-#define SDL_BUTTON_X2       5
+#define SDL_BUTTON_LEFT     uint8_t(1)
+#define SDL_BUTTON_MIDDLE   uint8_t(2)
+#define SDL_BUTTON_RIGHT    uint8_t(3)
+#define SDL_BUTTON_X1       uint8_t(4)
+#define SDL_BUTTON_X2       uint8_t(5)
 #define SDL_BUTTON_LMASK    SDL_BUTTON(SDL_BUTTON_LEFT)
 #define SDL_BUTTON_MMASK    SDL_BUTTON(SDL_BUTTON_MIDDLE)
 #define SDL_BUTTON_RMASK    SDL_BUTTON(SDL_BUTTON_RIGHT)
